+++ Systems/Enemies/DungeonBoss.cs
@@ -60,6 +60,7 @@ public class DungeonBoss : Enemy
             Defense = stats.Defense;
             XPValue = stats.XPValue;
             LootTable = new LootTable(minGold: stats.MinGold, maxGold: stats.MaxGold);
+            AsciiArt = stats.AsciiArt;
         }
         else
         {