+++ Systems/Enemies/Mimic.cs
@@ -35,6 +35,7 @@ public class Mimic : Enemy
             Defense = stats.Defense;
             XPValue = stats.XPValue;
             LootTable = new LootTable(minGold: stats.MinGold, maxGold: stats.MaxGold);
+            AsciiArt = stats.AsciiArt;
         }
         else
         {